Move dropdowns templates into separated directory
