+++ src/ggrc-client/js/components/dropdown/autocomplete-dropdown.js
@@ -3,7 +3,7 @@
     Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
 */
 
-import template from './autocomplete-dropdown.stache';
+import template from './templates/autocomplete-dropdown.stache';
 
 export default can.Component.extend({
   tag: 'autocomplete-dropdown',
