+++ src/ggrc-client/js/components/dropdown/templates/autocomplete-dropdown.stache
