+++ src/ggrc-client/js/components/dropdown/templates/dropdown-component.stache
