+++ src/ggrc-client/js/components/dropdown/templates/multiselect-dropdown.stache
